Updated openstack/openstack Project: openstack/swift  b3dd6a5df1ac4a5265b330787abcd52c7e116208
Improves log message in swift.obj.replicator

Before this commit, when a local device has not found been found
in a object-replication run, the policy was not mentioned in the
error log. But it is of interest to know the policy, for example for
error searching, when no local device has been found.

Change-Id: Icb9f9f1d4aec5c4a70dd8abdf5483d4816720418