Updated openstack/openstack Project: openstack/python-openstackclient  bf9b0e5ff4b56c73a603e9cd63565ff99a40ae83
Fix router set --route option

Fix the "--route" option on the "os router set" command. The
option did not properly format the new routes to set which
resulted in a "HttpException: Bad Request" error. In addition,
the output for routes was fixed to improve readability and to
align with the "--route" option on the "os router set" command.

Change-Id: I9c514153ec201e2feae32be6dd281771e3298b9c
Closes-Bug: #1564460