Updated openstack/openstack Project: openstack/nova-specs  31a6d202787f2542cd9d6cb4660a15febccf4129
Add in the Newton review priorities

Based on the notes from the Newton summit etherpad:

https://etherpad.openstack.org/p/newton-nova-summit-priorities

I've decided to omit non-priority stretch goals like
volume multi-attach since that is still a moving target.
The priority of that work can be re-assessed if we complete
some of the items in this list early enough to take on more
work before feature freeze, and it doesn't impact our other
priorities.

Change-Id: I4811b03eb0b38598c18708307890d8da95c0b285