Updated openstack/openstack Project: openstack/cinder-specs  e2d143ede4a74f6c329fc720df92dd72fcbde940
Remove Manager Local Locks for HA A/A

This spec outlines work needed to remove local locks from Volume nodes.

Two methods are used, filters on the compare-and-swap mechanism and lock
mechanism using workers table.

This is one of the aspects that are needed to support Active/Active in
cinder.

Implements: blueprint cinder-volume-active-active-support
Change-Id: Iab5bf8c3971d4a908bec16882fb4ca4a50c7dfdb