Updated openstack/openstack Project: openstack/horizon  bf294e4d63f9dbb85ae07cd23909d0c990bbf801
Improve Volume selection handling in LI

The code for updating the Volume Size error in the Launch Instance
source step appears to be incorrect. This patch updates and improves the
handling to also bump the current value to the minimum for the chosen
image (rather than just a form field error). Also changed the default to
create a new volume, as it was advised that this is the more common
workflow.

Change-Id: Iecb679d967b2285776278f23018e4151ebb18df2
Closes-Bug: 1568866