Updated openstack/openstack Project: openstack/ceilometer  800034dc0bbb9502893dedd9bcde7c170780c375
Add a tool to clean the legacy alarm tables

Since we have moved alarm out of ceilometer, when run ceilometer-dbsync,
the alarm tables will still be created and won't be removed. This change
added a tool that allow users to drop the legacy alarm and alarm_history
tables with SQL backends.

Change-Id: Ia831741319fd8e198f54a7d4a99b4d7833074e8e
Closes-Bug: #1578128
Signed-off-by: liusheng <liusheng@huawei.com>