Updated openstack/openstack Project: openstack/ironic  97d1fb6c12f5b8b56283ffebe20901474a26afe3
Fix API node name updates

This change fixes an issue when it is possible to specify name
multiple times in the patch, and only first of them is checked
for correctness. Instead, all of them should be checked, as stated
in JSON PATCH RFC https://tools.ietf.org/html/rfc6902.

Closes-Bug: #1566731
Change-Id: I4301876d9d4b060c69616b893dcdbc36e41a6369