Updated openstack/openstack Project: openstack-infra/project-config  934aa94854d7257a0021473af75c696fc44c8d56
Install constrainted version of babel

Instead of using a hardcoded version of Babel, let's use the constrained
version. This is still 2.2.0 right now but once it changes, it changes
here as well. We can make this change now since Babel is tested as part
of the requirements repository.

When installing horizon, use the same branch as the project uses
and use constraints here.

Test babel changes in requirements now using a tox target. The job is
only run when upper-constraints or one of the test files is changed -
and only run for master and newton or newer.

Change-Id: I15fa6a706323615d6072d8acaa859fea9bf35883
Depends-On: Iab7844ba748ecd84ec07ad4280b2317bd626fe9a