Updated openstack/openstack Project: openstack-dev/devstack  5899d20bc3aabdf543f99f8d06b8452175760df4
Set default value and configurable value for integration bridge

When using XenServer as hypervisor, install_os_domU.sh will create
integration bridge for compute node when neutron network is used.
But it should provide a way to allow moving of the VM to another
host (with a different XEN_INTEGRATION_BRIDGE) for easier install.
This patch is to provide the way to let user have the chance to
configure integration bridge themselves

Change-Id: If923a5e978e77fc091d24b6e1fe7a83a3375da09