Updated openstack/openstack Project: openstack/neutron  ad668e194144e51c0f27c6d1f6a322a4e4dffd1e
Delete post_test_hook.generate_test_logs

Log files as .txt files, don't zip them, and put them where
they need to be instead of copy them there in the post gate
hook. The benefit to doing this is that we'll get logs
for tests even if the job timed out.

Change-Id: I4bfd27534c827aed3cbd7b43d7d1289480ea4806
Related-Bug: #1567668