Updated openstack/openstack Project: openstack/governance  ca0c1824bb6b60aa1845bdd90a338ca67d48e92a
Add js-openstack-registry-hooks (openstack-registry-hooks) to openstack-infra

This project is a set of event hooks, which assist in managing the
lifecycle of openstack-infra's NPM mirror. At this time, two hooks are
implemented: One which exits the registry process as soon as the mirror
has been sychronized (so we can run vos release), and a second that
sanitizes data from the skimdb, to make sure that it matches what's in
the public registry.

Change-Id: I55f5dc0ee7ecaf64c59d6b8f0a7a4c1ff96db2b5
Depends-on: Ib1e5ac2ea54cceb0d2c853ed134f355396debeb6