Updated openstack/openstack Project: openstack/neutron  b1b346f81100d8c813621a76452b525a383dd962
DVR: Fix TypeError in arp update with allowed_address_pairs

There is a TyperError seen when trying to update an arp entry
for the DVR service port that has associated allowed address
pairs.

The _get_allowed_address_pair_fixed_ips were not returing the
fixed_ips dictionary but just returning the fixed_ip string.

This caused the TypeError to occur. In addition to the fixed_ip
it should also return the subnet_id for arp_update and it was
returning it.

This patch fixes the issues mentioned above.

Closes-Bug: #1566046
Change-Id: I35d3145682e23e087c17129bb9556946e0aa801c