Updated openstack/openstack Project: openstack-infra/project-config  07f78a1a8b6c684675c9d93d1338e05b36cb962d
Disable App Catalog periodic job

This commit disables the periodic job that checks for dead links and
submits a proposal with an update if any dead links are found. The job
is having trouble following some links that have redirects which
results in live links being flagged as dead. We will also adjust this
in the near future to take direct action rather than proposing a
change via gerrit.

Change-Id: I97f012eee4b23bf8deed765a7db73caaa1dc326d