Updated openstack/openstack Project: openstack/nova  ef43123fdded2de5386db904e0e98b2895b8bdf8
Fix unit tests for v2.1 API

VolumeApiTestV21 is a test class for v2.1 API as the name, but the
class tested v2.0 API because the module of v2.0 API was used.
In addition, the specified request bodies also were wrong because
the parameter "size" is required on JSON-Schema.
This patch fixes these problems.

In addition, this patch renames an internal method of the related
test because the method name was difficult to understand the purpose.

Change-Id: Ibe165f00eb2e15a1a2d9466c1af48db2e6883916