Updated openstack/openstack Project: openstack/neutron  f18db6f754b36e1960796ee883efac78f0d44007
DVR: Use existing IPDevice to add address on FIP VETH

create_rtr_2_fip_link() has IPDevice instances for each
end of the VETH pair it created, just use them to add the
link-local IP to each one.  Because this is a point-to-point
link, we don't add the broadcast address when we assign the IP,
so add a flag to IpAddrCommand() class to accomodate us.
This is also setup for a follow-on patch.

Partial-Bug: #1566383
Change-Id: I66d70afaf3c52861e0887d3f535d5327ea46b128