Updated openstack/openstack Project: openstack/cinder  8091e9f7374aa251414214364b9bd6540592c1da
VMware: Support for paraVirtual image adapter type

Currently the VMDK driver does not support creating volumes
with paraVirtual adapter type. It fails the copy image to
volume operation if the image's 'vmware_adaptertype' property
is set to 'paraVirtual'. This patch adds the paraVirtual
adapter type to the list of valid adapter types to fix this.

Closes-bug: #1578399
Change-Id: I981b1737974ae1fbb3eb3fef3811bcb50030d670