Updated openstack/openstack Project: openstack/neutron  36a04c9a308ec498f8e85b439a3b95ae22a15092
Partial revert "DVR: Fix issue of SNAT rule for DVR with floating ip"

Commit 1cea77b0aafbada6cad89a6fe0f5450004aef4e1 try to resolve 2 problems

1) The prevent snat rule that is added for floating ip will be cleaned,
when restart the l3 agent.
2) Ping floating ip, but reply with fixed ip

To resolve 2), SNAT to shared gateway was introduced for internal traffic
to floating ip inside the qrouter namespace. But there is a potential issue
when multiple router SNAT to one address. The destination VM might not be
able to differentiate the incommint request. To avoid the potential issue,
revert the code for 2), but reserve the code for 1) in this patch

Change-Id: I5d11005d0aca681af99a011812ea53f92f2a400d
Co-Authored-By: Hong Hui Xiao <xiaohhui@cn.ibm.com>