Updated openstack/openstack Project: openstack/designate  281b4f9f2e20f5faf9888d16bc79bae5373c101e
Fix parameterized class decorator

The decorator use the types.MethodType to attach the parameterized
test to the test class, but the function changed in python 3.x. This
uses six to do the right thing so this can be used outside the tempest
tests that are always run with python 2.x.

Change-Id: I56fee078d3fa718536208fd3bd5228cc45d24444