Updated openstack/openstack Project: openstack-infra/tripleo-ci  4ce3652e0893aaab8672a1513b5998c3139ad095
Calculate deployment times for the Heat resources

Looking through our graphite metrics, by far the longest time is
spent deploying the overcloud.  This change allows us to collect
more granular data about what is using that time in the deployment.

At some point we may want to add this to the graphite data so we
can track it over time, but just making it available in the postci
logs is a good first step.

Change-Id: Ibae0f6259146dc0da73d43aba0ff31a24ae71eff