Updated openstack/openstack Project: openstack/nova  0f27c16b46be7569bd9a67a15edefe0a287949ad
Fix database poison warnings, part 7

The following warning appears in the unit test logs a number of times.
This patch fixes all the warnings in unit.compute.test_compute_mgr.

    "UserWarning: This test uses methods that set internal oslo_db
state, but it does not claim to use the database. This will conflict
with the setup of tests that do use the database and cause failures
later."

Note that this warning is only emitted once per unit test worker, so new
offenders will show up in the logs each time you fix a test until they
are all gone.

Change-Id: Ie9a3f3a1d5f38fb81fff37922753dfca10b96531
Related-Bug: #1568414