Updated openstack/openstack Project: openstack-dev/devstack  ea15ca1318825fb01b1a077b04132bc2912ce7be
Adjust region name for images, networks and volumes during deployment

In a multiregion installation of devstack, conflicts occur in the
creation of images, networks, or volume types, when the region is not
set.

This patch adjusts commands to include the region, and
also adjusts the region_name in the nova configuration section of
neutron.conf to include the region.

Change-Id: Ifedff6a124fa49d57cc7b2f35916d8d96f5e5f7a