Updated openstack/openstack Project: openstack-dev/devstack  251b870e859e1a03b57c8fe24c7692a5e2a10fde
Update auth params in Nova Hypervisor-Ironic

As Nova hypervisor uses deprecated parameters when trying to
authenticate to Ironic, as well as a hardcoded /v2.0 endpoint, a fatal
error occurs when creating a keystone v3-only devstack.

This patch updates auth parameters (ironic section in nova.conf) that Nova
uses when trying to connect to Ironic to v3 parameters.

Change-Id: I2d7ebf750115613aa917448f20daaece614633ef