Updated openstack/openstack Project: openstack-infra/project-config  62ab336db7957136a780c12daa618fd06edf57d7
Improve pot publishing

Some projects like horizon have several modules and each contains the
same pot file, just in different locations:
* horizon/locale/djangojs.pot
* horizon/locale/django.pot
* openstack_dashboard/locale/djangojs.pot
* openstack_dashboard/locale/django.pot

so, the previous logic copied all in the exact same directory, just
resulting only in one of each - and not both.

Change the logic to use exactly same path for publishing. This is also
the path that is shown for each file in Zanata, so it's easy to
recognize and consistent accross git repository, Zanata and
tarballs.openstack.org.
Instead of a final find, run find only in the modules that we handle.

So, files are published to:
tarballs.openstack.org/translation-source/horizon/master/horizon/locale/djangojs.pot
etc

Change-Id: I2b5f6383b2d918ffdbbf4adc1dcccd06d721c65f