Updated openstack/openstack Project: openstack/ironic  9996949e8e056ae6c434ec9e3708fe49fa794441
API: Check for reserved words when naming a node

This patch is introducing a mechanism which prevents nodes to be named
with some reserved words that are implicitly not allowed today due the
way the Ironic API works.

For example, the way we fetch a node by its name in the API is issuing
a GET to the http://<address>/v1/nodes/<node name> but, we have some
words that are valid endpoints under the same path. One example is of it
is the /detail endpoint, a GET request to v1/nodes/detail should return
detailed information about all the nodes.

Apart from that, due the way pecan/wsme parses/routes the requests
the words 'ports', 'vendor_passthru', 'management', 'maintenance' can
not be used. These words are attributes pointing to a controller in the
node's API object (e.g v1/nodes/(ident)/ports) and naming a node as such
confuses pecan/wsme.

The API microversion was not bumped because this error should be
prevented across all versions.

Closes-Bug: #1572651
Change-Id: Ibba5ed16e6961805864bdf46b94296b1b0c09469