Updated openstack/openstack Project: openstack/nova  31972f8cbe722160abc841d0d48a49fe8812a77e
Fix database poison warnings, part 8

The following warning appears in the unit test logs a number of times.
This patch fixes the warnings in virt.libvirt.test_driver that come from
driver._get_connection() calls, but there are still a bunch more
warnings in that file.

    "UserWarning: This test uses methods that set internal oslo_db
state, but it does not claim to use the database. This will conflict
with the setup of tests that do use the database and cause failures
later."

Note that this warning is only emitted once per unit test worker, so new
offenders will show up in the logs each time you fix a test until they
are all gone.

Change-Id: Icacce09f8a83cb4c73d5b6caf5cd31b55032c847
Related-Bug: #1568414