Updated openstack/openstack Project: openstack/keystone  4304c17713794c8ad74ec826a371536349efefd6
Move the resource abstract base class out of core

This patch moves the resource abstract base class and other backend
code out of core and into backends/base.py

This removes dependencies where backend code references code in the
core. The reasoning being that the core should know about the backend
interface, but the backends should not know anything about the core
(separation of concerns). And part of the risk here is a potential for
circular dependencies.

Partial-Bug: #1563101

Change-Id: Id14d698fa7e083e72c463732e227726448a86f6d