Updated openstack/openstack Project: openstack/neutron-vpnaas  a65310b7bbc1e48382876c8ac4868e60a0d47ba7
Add grenade upgrade hooks to stop agent

This adds grenade upgrade hooks so the agent can actually
be stopped during an upgrade.

This is necessary now that vpnaas runs in a devstack plugin
because its agent will now no longer be stopped by the neutron
plugin inside of grenade.

Change-Id: I5271f0ef4f857eb5650612829d352a821416487a
Partial-Bug: #1579235