Updated openstack/openstack Project: openstack/oslosphinx  75cfb9f93b3560ff412cae4de3029157c14b258f
add recent versions links to sidebar

The last five tags (that start with a number) are used as version
numbers and linked to. This updates the sphinx extension so that
every user of the theme gets this info automatically with no need
to update the conf.py on a per-project basis.

Change-Id: I9bf17872bd23bed3cab9b5e2349d615c7001cb66