Updated openstack/openstack Project: openstack/glance  da31cfff614b46aaf68edf25daf85b9b464f449d
Return BadRequest for 4 byte unicode characters

Various create and update apis are returning 500 error
when you pass 4 byte unicode characters in the request.

This patch fixes the issue to return HTTP 400 Bad Request.

1. md-namespace-create
2. md-namespace-update
3. md-property-create
4. md-property-update
5. md-object-create
6. md-object-update
7. md-tag-create
8. md-tag-update

APIImpact: For all of the above APIs, it will return 400 error instead
of 500 if user passes 4 byte unicode characters in the request.

Closes-Bug: #1570789
Change-Id:  I8233cf95310aa2671ebc62995c58208024fc38a3