Updated openstack/openstack Project: openstack-infra/project-config  38d021a58adf8eeb7fe16cf76dd03fd68e390396
Adding tempest-horizon repo and job

In an effort to make the overall gate more robust and still
allow Horizon to update their xstatic packages, this patch
adds a repo containing the current horizon tempest scenario
tests in tempest plugin form. The patch also adds a non-voting
check job to allow testing of the plugin.

Without this change, it is possible to break the gate by
updating xstatic packages before Horizon can support the new
version. This sad truth is an offshoot of the difficultly of
supporting the many incompatibilities between versions of
javascript libraries.

Once the new job is verified and added to the horizon gate, the
current horizon scenario tests in openstack/tempest would be
removed.

Needed-By: Ib01a69ffb978708ad7b515d3aaa4b58bc894a167
Change-Id: I1608407a4f8246728c30870fd5e1886c632e7734