Updated openstack/openstack Project: openstack-infra/irc-meetings  7eb36153e00f999376e15394dc4658619276b3d1
Change time and irc channel of Vitrage project weekly meeting

I have checked all the meetings at that time by importing the irc
file of the meetings to my calendar, and saw that this channel is
free at that time.

For clarification, Vitrage already has a weekly irc meeting and we
just wanted to change the time.

Change-Id: Ib360c1b526482ee5ca600b266ed91700e88597bd