Updated openstack/openstack Project: openstack/nova  f5b4f047b1ad25c9565030fd72c2bc25645a794e
pci: pass in instance PCI requests to claim

Removes the calls to InstancePCIRequests.get_XXX() from within the
claims.Claim and claims.MoveClaim constructors and instead has the
resource tracker construct the PCI requests and pass them into the
constructor.

This allows us to remove the needlessly duplicative _test_pci() method
in claims.MoveClaim and will allow the next patch in the series to
remove the call in nova.pci.manager.PciDevTracker.claim_instance() that
re-fetches PCI requests for the supplied instance.

Change-Id: Ib2cc7c985839fbf88b5e6e437c4b395ab484b1b6