Updated openstack/openstack Project: openstack/nova  ab952e0c287ad67c2e4b554c2cb224fb1aacb3f3
VMware: create temp parent directory when booting sparse image

When sparse images are uploaded with HTTP PUT to a vmware datastore,
the parent directory may not exist and it will be automatically created.
However, this is not the case when using datastore copy for fetching the
image. In this case we should create the parent directory first.

Change-Id: I18783e23ba487504e9130171c7424b807ad0251d
Closes-Bug: #1552610